requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
google-genai>=1.55.0
pydantic>=2.7.0
python-dotenv>=1.0.0
httpx>=0.27.0
certifi>=2024.2.2
PyJWT>=2.10.0
cryptography>=44.0.0
